NSUInteger for index
git-svn-id: svn+ssh://svn.savannah.nongnu.org/gap/trunk@1789 d25b4930-3292-e311-b074-0010c6de5da8

+++ system-apps/Addresses/Test/Controller.m
@@ -1,8 +1,8 @@
 /* All Rights reserved */
 
-#include <AppKit/AppKit.h>
-#include <Addresses/Addresses.h>
-#include <AddressView/ADSinglePropertyView.h>
+#import <AppKit/AppKit.h>
+#import <Addresses/Addresses.h>
+#import <AddressView/ADSinglePropertyView.h>
 #include "Controller.h"
 
 @implementation Controller
@@ -16,7 +16,7 @@
   [propSelector selectItemWithTitle: [propView displayedProperty]];
   [propView setDelegate: self];
 
-  int index = [autoselPopup indexOfItemWithTag: [propView autoselectMode]];
+  NSUInteger index = [autoselPopup indexOfItemWithTag: [propView autoselectMode]];
   if(index != NSNotFound)
     [autoselPopup selectItemAtIndex: index];
   else
